requirements.txt
streamlit==1.35.0
tensorflow==2.13.0
numpy==1.24.3
pillow==10.3.0
